+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - improve: select all text when edit text
 - fix: some workshop links can't be parsed
 - improve: auto refresh images after downloading
+- fix: add try/catch block on opening clipboard
 # v0.4.0
 - fix: renaming bug
 - improve: change the deletion behavior from direct deletion to move to recycle bin (Windows only)